Fix bug in debug build caused by aggressive check
In the case of a function with 1 argument, the evaluation of the
argument is done into a single cell with internal termination and
garbage collection that lives in the Reb_Frame structure itself.  This
saves on needing to push a chunk of values.  However, this means that
internal cell is not available to the 1-arg routine as scratch space
while it is running.

A previous commit enhanced the debug checking to catch 1-arg routines
trying to use the scratch space.  It also made it so 0-arg routines
*could* use the scratch space, by not pointing it at the cell, but
using a garbage value to catch those who tried to enumerate args when
there were none.

This is a bug, because the way args are enumerated when there are none
is to look for the END_CELL terminator.  So one can't put garbage in
the arg slot even if it's not "being used" to make enumeration not
crash.  Hence this commit uses the global END_CELL as opposed to a
0xDECAFBAD pointer.

As optimization goes, this allows 0-arg routines to use the GC-protected
scratch cell which is already available.  But 0-arg routines aren't
that common, so the extra logic to switch on the count of arguments
on each evaluation may be worse than just sacrificing the cell and
running the same code path as the 1-arg case.  It's likely more
useful for its debug check properties...e.g. preventing the accidental
use of the scratch cell from suddenly making an argument appear that
gets enumerated, making a 0-arg routine look like it has 1 arg.