SERIES_FLAG_ACCESSIBLE => SERIES_FLAG_INACCESSIBLE
The sense of the flag for detecting inaccessibility is backwards for
doing a generic check, because the default when the header of a
REBSER is initialized is to have the bits at 0.  Since most series
are accessible and only the rare ones are not, it's more efficient
to switch the sense of the flag.

As long as this was being changed, also switches the native that
makes a struct inaccessible to use a variadic invocation of Do, which
is less code and overhead than constructing an array.

Also gives clearer names to the mark flag routines that assert the
prior state of the mark.