Update %m-gc.c code and comments, remove all C stack recursions
This is a thorough going-over of the mark-and-sweep garbage collector
code, with an eye toward catching the edge cases and simplification.
Several changes likely improve the performance some amount, by
avoiding redundant markings or checks.  However, the big push was
to reduce the total amount of code and improve the quality of the
comments and function names.

One vetting that this does is to completely eliminate C stack recursion
in the GC.  Deeply nested structures in R3-Alpha could overflow the
C call stack with Mark_Array calling Mark_Array calling Mark_Array...
but that was solved with a queue in Ren-C.  However, that wasn't the
only possible problem, just the most likely.

With additions like objects having meta objects, if you call
Queue_Mark_Context of CTX_META(context) from inside Queue_Mark_Context
you could theoretically get the same problem.  The same goes for
marking underlying functions while marking a function.  While neither
of these are too likely to make chains too long for the C stack, it
is possible--and the discipline of organizing the code to stop this
kind of problem helps for any similar patterns which *would* be
likely.