Bring back the REBVAL tunneling in FFI
While the ordinary types used to pass to the FFI are varied sizes of
C types (8-bit unsigned integers, 32-bit signed integers, pointers)
the FFI had an extension allowing a parameter to be marked as
"REBVAL".  This allowed the FFI to be passed a full sizeof(REBVAL)
cell as a parameter to routines.  Though it's not likely a library
you would have to call through FFI would be linked to Rebol in any
way to know what to do with a REBVAL, it could be interesting if the
value were to be tunneled through a call *back into Rebol* as an
argument.

The feature created complications for the implementation changes making
the FFI use ordinary Rebol structures (instead of C structs) for its
implementation.  This reimplements it, though rather than passing the
REBVAL as a full cell it uses a pointer.  Since the focused application
of this is for tunneling a parameter from a routine into a callback
that is running while the routine is still on the stack, the lifetime
of the value should be safe.

In the first cut of the implementation, after parsing a struct the
field types were encoded as integers from libffi's FFI_TYPE enum.  That
enum didn't have an entry for FFI_TYPE_REBVAL (obviously) so there had
been an "is_rebval" flag, which told the Rebol FFI layer that although
the lower-level FFI was in terms of pointers for a REBVAL, parameter
proxying needed to be handled differently from a plain pointer.
(e.g. an INTEGER! being passed to an ffi pointer field actually passes
the integer's value as a pointer, while for a REBVAL it wants the
integer *cell* address).  However, it's clearer for readability to
have the struct speak in terms of words, so SYM_REBVAL and SYM_UINT8.

(Eventually, it may be that structure schemas will be fairly ordinary
looking objects, which do not need any processing to be clearly read
by the user.)