Switch hardcoded integer indexes to enum for RIN/FLD
When drafting the implementation of using Rebol arrays for struct
fields and routines, it was expedient to use integer indices (they
were only really assigned once, and the read routines used after that).

With the implementation working, it's good practice to go through and
get rid of the magic numbers.  (Also, it turned out there were several
assignment points in the struct code.)

This adds index enumerations and does a little reorganization.