Val_Init_Object => Init_Object, etc.
Originally Ren-C followed along with Rebol's VAL_INIT_XXX naming
scheme, in particular because there was some ambiguity between some
of the series types and value types.  So "Init_Block" wasn't obviously
speaking about a REBVAL, it could be speaking about a BLK (an
ANY-BLOCK! series, what is now known as an ANY-ARRAY! and subtyped
as a REBARR)  With that ambiguity gone, Val_Init really just makes
things less readable.

This makes the changes for contexts (errors, ports, objects...)