Separate GC marking flag from recursion marker bit, and fix bug
R3-Alpha re-used the same marking flag from the GC in order to do
various other bit-twiddling tasks when the GC wasn't running.  This is
an unusually dangerous thing to be doing...because leaving a stray mark
on during some other traversal could lead the GC to think it had marked
things reachable from that series when it had not--thus freeing
something that was still in use.

A bug inadvertently introduced in Ren-C when adapting the PROTECT code
(some months ago) made this mistake, and it was difficult to find.  The
issue arises when literal objects occur in blocks that are marked deep,
so something like:

    protect/deep compose/deep [[[(make object! [ae: 1020])]]]

While leaving a stray mark on is a bug either way, when a bug triggers
GC problems they are particularly hard to track down.  So one doesn't
want to risk them if not absolutely necessary.

(Not to mention that sharing state with the GC that you can only use
when it's not running gets in the way of things like background
garbage collection, etc.)

This commit splits out the GC bit from the generic bit.  It keeps the
term "mark" for the GC, since that's standard nomenclature.  A lot of
basic words are taken other places for other things (tags, flags)
so this just goes with a series "color" of black or white, with white
as the default.  The debug build keeps a count of how many black series
there are and asserts it's 0 by the time each evaluation ends, to
ensure balance--which is inexpensive and a good safety net.