Initial Implementation of Immutable Source
When specific binding was implemented, function bodies became
immutable.  This was because the same body block was being "invisibly"
shared among recursions of a function, even though words in it were
being viewed as having a different binding.  If the body were allowed
to be changed by one recursion, the other bodies which were effectively
their own unique blocks of code, would also be changed.  While a
generally dangerous idea to begin with, it became moreso because any
references to locals or arguments would be "specific" yet then
inserted into a relativized body which would have meaning only for
that execution frame.

That was a fairly loose form of immutability, because it used R3-Alpha's
PROTECT bit.  That bit was user-owned, and could be cleared.  This
strengthens immutability in the system by making 3 distinct bits:

* PROTECTED - the user-controllable protection bit that can be added
  shallowly or deeply, and removed at any time.  It's more of a debug
  feature or convenience than anything.

* RUNNING - this is a transient state of read-only-ness which lasts
  as long as a series is being held onto by the evaluator.  This
  prevents things like `code: [insert code 'print] | do code`.  That
  is a questionable feature to begin with--and more likely a bug.
  But moreover, allowing such mutation means that DO has to be
  somewhat "crash-proofed" because pointers can move out from under
  it.  In fact, such modifications could fully invalidate the pointer
  to the block if it had to expand.  The difficulty of providing a
  useful behavior semantic combined with the runtime cost of crash
  proofing makes protecting the series a better bet.  The user cannot
  disable this protection, but it's released on completion/FAIL/THROW

* FROZEN - This is the new form of immutability that source code gets.
  The user may request any series to be frozen with the new primitive
  LOCK, but unlike with PROTECT/UNPROTECT there is no UNLOCK.

Source freezing is done whenever code is loaded "as a script".  This
does not apply to LOAD of a STRING!, so it's legal to write
`do load "append {abc} {de}"`.  However, if that code appeared in a
file or module then the append would be an error.

A new requirement is that any series which is to be used as a key in
a map must be locked.  This addresses a longstanding bug in R3-Alpha
where strings inserted as keys in maps could be modified without
updating the hash, and strengthens the feature of Ren-C of being
able to allow blocks and other arrays as hash keys--so long as they
are locked.  Since literals in source qualify, many values will be
"pre-approved" to act as map keys.