JOIN => JOIN-OF, reserve JOIN, act like R3-Alpha REPEND
The classic definition of REDUCE doesn't necessarily REDUCE in
R3-Alpha:

    >> x: 10
    >> reduce 'x
    == x

This was remedied in Ren-C and in Red.  The resulting consequence
was that REPEND also began reducing things, as its definition was
REPEND SERIES VALUE => REPEND SERIES REDUCE VALUE.  However, this
may or may not be as useful for REPEND's common case.

In fact, there is another word in Rebol space that one could
think of as being more appropriate as being a variant of what REPEND
is doing--namely JOIN.  Historically, JOIN is just a non-modifying
version of REPEND.  Its name does not have the baggage of sensibly
needing to mean "apPEND buffer REduce x"...it can take artistic
license and do whatever it's most useful for a JOIN to do.

Yet in the Rebol world, basic terms (APPEND, INSERT, REMOVE, UPPERCASE)
are modifying by default.  Ren-C is trying to bring in a vocabulary for
immutable variations of these, which could be optimized via a
refinement (e.g. `APPEND/COPY X Y` can be made more efficient than
`APPEND COPY X Y`, because APPEND can know the length of the final
result ahead of time from the sum of the lengths of X and Y, and do
one allocation for that size, vs COPY doing one allocation at X size
and then APPEND needing to expand that.)

So it makes sense to give the mutable operation to the undecorated
name, JOIN.  And for the immutable version (a.k.a.) JOIN/COPY, this
can be abbreviated to JOIN-OF.  (Abbreviation has the advantage that
it can be done as `join-of: adapt 'join [series: copy series]`,
even before the refinement is implemented.)

And it looks better, also making a better impression on new users than
REPEND does:

    buffer: copy ""
    emit: func [x] [join buffer x]

This goes ahead and refactors REPEND, JOIN, and JOIN-OF...as well as
reframing REJOIN to be in terms of JOIN.  It also makes REJOIN []
return BLANK!, as opposed to a BLOCK!, since it has no basis for
knowing what content type to return.

In order to let JOIN-OF have time to settle, a synonym for the new
JOIN is established as ADJOIN and JOIN gives an error message.