Get rid of EQUIV? tests
In light of the number of issues to sort out regarding basic
comparison, this removes the confusingly and unattractively named
EQUIV?.  The original goal of the operation was to tell whether two
words shared not only the same spelling but also the same binding,
but further to be able to apply this question to entire blocks of
code.

The question about words can be answered currently by SAME?, since
words are only considered the SAME? if they have the same spelling
and same binding.  Other nuances of the equality comparison are not
worth the weight of the large number of redundant tests.

In any case, EQUIV?'s implementation was deleted very early on in
Ren-C, but the tests weren't.  Finally got tired of seeing the
pointless tests go by.  :-/