Fix 0-arg case for args_head
To help with debug checks of 0-arg functions, instead of using a
writable cell with an END marker in it, it was switched to use the
read-only END_CELL.  But there were two cases that needed to check
for that in the freeing case, and one didn't get accounted for.
This adds the END_CELL check.

(Note: As was observed in that commit, checks like this probably end
up making it not an optimization, and it is of questionable use to
free up the D_CELL on 0-arg functions.  But it's a good debug check,
so it should probably be converted to debug only.)