Use newly freed bit to mark REBARRs as map pairlists
The GC has moved toward array subclasses "knowing what they are", so
that if there is a pointer to any REBARR it can participate in GC
and manage the ->link and ->extra fields (or any other knowledge).

Currently the bits are not compacted into a "type" so that masking is
not necessary.  But since it's looked at by assertions most of the
time they probably will be turned into an ARRAY_TYPE eventually.