Tuning of the GC scheduling
Include REBSER and REBIN in GC_Ballast calculation.

The way the GC is scheduled is like:
GC_Ballast starts with MEM_BALLAST (3000000 by default), and reduces
when memory gets used, and increases when the memory is freed, when it
reaches zero, a GC is scheduled. After GC, GC_Ballast is set back to
Task->Ballast (same as MEM_BALLAST, if user didn't changed it).

Because the memory pools never shrink, a less frequent GC could cause
the memory usage to rise.

Include every memory usage in the pool should make the calculation more
accurate.