Simplify GC disablement, allow GC active during boot
In R3-Alpha, disabling the garbage collector was done somewhat often,
because there was no such thing as the tracking of "manual" series
that had not become "managed" yet.  Ren-C removed the necessity for
this, so that the GC could always stay running.

One exception to this was the boot process, which couldn't have the
GC running--even after its basic setup.  This meant the first execution
of Recycle() was around the 9000th DO step in the evaluator.  That
meant if there was a problem with the GC, it might be buried early
in the boot process and not discovered on the actual step that created
the problem (which ordinarily could be found by "torturing" the GC,
running a full Recycle on every step).

That problem was encountered when trying to add `prior_specifier` to
contexts.  To hunt down the evaluator tick in the boot where the
problem occurred, this went ahead and made it safe to run Recycle()
from tick 0.

This also narrows the handling of GC disablement to the few debug
routines that use it (to avoid complicating output during an
investigatory debug spew), and the user-exposed RECYCLE in a clearer
way...by splitting it into the user flag and the internal avoidance
of re-entry flag.