COPY_VALUE() => Derelativize()
As specific binding has settled, the question of whether or not a
REBVAL can be bit-copied without a function hook has arisen.  Several
proposals would involve code to run when one REBVAL* is assigned to
another, so that simply writing `*dest = *src` would not be sufficient.

While prohibiting such assignments and requiring a Copy_Value() routine
for REBVAL* has not yet been deemed a requirement, the name
"COPY_VALUE" for derelativizing REBVALs is questionable.  This changes
that to be more accurate, and leave room for something that isn't a
pure bit-copy for duplicating REBVALs in the future.