JOIN-OF uses FORM in bootstrap for WORD! series
In the new formulation of JOIN and JOIN-OF, JOIN is a mutating
operation on series...while JOIN-OF is the non-mutating version.  This
means that JOIN does not make sense on WORD!s (today) because WORD!s
are immutable.  Hence, JOIN-OF would also not make sense on WORD!s
(if the only definition of JOIN-OF is to copy its target series before
the joining).

The bootstrap scripts had a line where a word was JOIN-OF'd, which was
generating an error.  This pushes the responsibility to the caller to
FORM that word prior to the joining, which can be a JOIN since the
FORM'd series is mutable.  (ADJOIN for legacy reasons)