Make release build re-use most recently freed node
In R3-Alpha, the most recently freed node would become the first
node to hand out.  This is a simple and likely good strategy for
cache usage, but makes the "poisoning" nearly useless.

@ShixinZeng added code to insert an empty segment, such that the most
recently freed node would not be picked up by the next Make_Node.
Although that elongated the amount of time the node would stay
poisoned, this may have an adverse effect on performance/caching.

This puts freed nodes at the head of the freelist for release builds,
as in R3-Alpha.