Kill GETVAR_UNBOUND_OK, TRY_GET_OPT_VAR, TRY_GET_MUTABLE_VAR
The idea of fetching a variable and allowing for failure was part of
Get_Var_Core, so that if given the GETVAR_UNBOUND_OK flag it would
be willing to return NULL instead of fail()

It would be difficult to reproduce the exact mechanics of failure of
Get_Var_Core in client code.  However, the number of cases where a
meaningful behavior can be given besides erroring in client code is
rare, and many of the legacy clauses which were using these were just
planning to deliver an error anyway (probably with less locality to
the problem in question).

The most common case where code can likely do something meaningful
is in testing for unbound cases--and even then it is questionable.
(For instance, type spec blocks are willing to use a binding, such
as `i!: integer! | func [i [i!]] [...]` but are also willing to
go by the word-to-type mapping if there's an error, but is that a
good fallback if something like **print integer!** has a lookup
problem?)

For the sake of simplification, this moves the test for IS_WORD_BOUND
into the callers who think they can do something meaningful if the
lookup fails.  PUSH_TRAP can be used for those few/rare cases that
arbitrary failures can be handled.