Don't allow setting or getting generic value flags on ENDs
The GET_VAL_FLAG and SET_VAL_FLAG routines were low-level enough that
they were willing to set flags on "trash" values, and were not noticing
when values that were marked as END were having flags read off of them.
Since ENDs can be implicit, this was bypassing the ordinary blocks that
would prevent non-value-headers from being updated with value flags
(e.g. series info bits)

When some implicit END headers had their bit patterns changed, this
change in pattern caused the bit in the VALUE_FLAG_LINE slot to be
always set to true.  That raised attention to the fact that the line
flags was being both set (in %l-scan.c) and checked (in %s-mold.c) on
END values.

While relatively benign in this case, it could lead to serious errors.
This adds an assert to stop such sets or clears, and corrects the
code for the line ending markers.