REBSER*, REBVAL*, and UTF-8 char* distinguishable as void*
This commit is believed to finish (modulo bugs) the work on organizing
the header bits so that valid REBSER*, REBVAL*, and UTF-8 char*
sequences can be discerned from each other.

Any UTF-8 string whose first 2 bits are `10` is invalid.  That
sequence is reserved for "continuation bytes", and could never start
a string:

https://en.wikipedia.org/wiki/UTF-8#Codepage_layout

A valid REBSER* will always have `100` as its first three bits.  (A
freed series--which is invalid--will have `00000000` as its first 8
bits...which while not technically distinguishable from a UTF-8 empty
string, the debug build can heuristically guess whether the context
is more likely an empty string or freed node for debug output.)

A valid non-END marker REBVAL* will always have `101` as its first
three bits.

END markers which are full cells have `11111111` as the first 8 bits
(a byte which is never allowed in any UTF-8 string).  END markers which
are "implicit terminators" have `x10` as the first 3 bits; which is not
technically enough information to tell it is not a valid first byte of
a UTF-8 string.  The second form of END marker is for internal array
enumeration only, and will never be exposed via the API.