Count *down* n in Try_Find_Containing_Series
The debug routine for trying to locate the containing series of a
REBVAL was counting *up* from the number of units in the pool
instead of down.

This probably went unnoticed because it's a debug routine that
generates a useful stack trace either way.  Hence it wasn't easy to
notice that an intended extra diagnostic feature was missing.