Correct GET_VAL_FLAG() on END in -O2/64-bit/debug builds
The particular combination of -O2 and debug builds on 64-bit platforms
somehow exposes a case of GET_VAL_FLAG() on END markers that was not
appearing otherwise.  Basically the THROWN() flag is checked on an
END marker.

(It may be completely invalid, because the THROWN() bit is not
guaranteed to be legitimate on an implicit END marker.)

This corrects the problem by first testing for END before testing
THROWN().