Universal panic()!
This commit takes advantage of the ability to discern a UTF-8 string
from a REBVAL* or a REBSER* subclass to make it possible to use a
single convenient `panic ()` macro to request diagnostics Rebol
values, or deliver the panic message if there is no smoking gun.

It also involves a reimagining of the role of `panic()`.  R3-Alpha's
error machinery had a separate "boot string" table for crash messages,
while Ren-C tried to unify the messages into the same table as used
in %errors.r -- to provide a central point for internationalization of
both categories.  But as time wore on, the messages in a panic were
of less importance than getting good diagnostics--such as those
returned by Panic_Series under Address Sanitizer.

So this pushes away from the unification with the error machinery or
potential internationalization of messages about things the user
should never see (and could probably do nothing about).  Instead,
panic moves toward supplying the precise file and line number--even
if the panic appears in a release build, so that the argument to
panic can be something that might offer an actual clue about what
went wrong.

So instead of making an error like this:

    bad-boot-result: {There was a bad result during the boot}

And then writing something like this in the C:

    if (!IS_VOID(result))
        panic (Error (RE_BAD_BOOT_RESULT));

... or trying to inject some kind of assistance like:

    if (!IS_VOID(result)) {
    #if !defined(NDEBUG)
        printf("The boot result type number was %d", VAL_TYPE(result));
    #endif
        panic (Error (RE_BAD_BOOT_RESULT));
    }

The simpler thing is to not make an error and just say:

    if (!IS_VOID(result))
        panic (result);

This makes the code more readable, lets panic() report as much about
the result as it can reasonably think of in the log (not just the
value's type, but when it was initialized...if it lives in an array
then what evaluator tick it was created on, etc.)  That plus the
file and line number is--for a panic--much more likely to provide
good information "in the field".

But if no "smoking gun" is available, you can just use a message.

    panic ("this will work too")