Eliminate redundant code by unifying GUARD_VALUE and GUARD_SERIES
It has been possible to discern REBSER and REBVAL for a while, since
they both adopted the same Reb_Header format.  However, this was an
emergent feature which wasn't necessarily taken for granted as a
"design requirement".

Several system changes have made it essentially a requirement to be
able to tell the difference--most notably the handling of individual
API-exposed GC'able REBVALs that live in REBSER nodes as "pairings".
More features are depending on this design choice, and given that it
is redundant to have two identical bodies of code for guarding REBVAL*
and REBSER* respectively.