Bring some heavy checking #ifdef'd code up to date
Some checks are too expensive to run with generally, but are good to
run on occasion (and would be good to include in the Travis CI build
matrix, at least to run occasionally).

This brings a few checks up to date, such as STRICT_BOOL_COMPILER_TEST.
That test is what ensures REBOOL always contains a 0 or 1, to avoid
situations like:

    void SomeFunction(REBOOL flag) {...}

    if (bitmask & thing)
       { ... treat as "truthy" ... }

    SomeFunction(bitmask & thing); // oops... flag is not 0 or 1

By guarding against such mistakes, optimizations can take advantage of
the zero-ness or one-ness, e.g. by using to shift or bit mask or
casting the REBOOL to an integer that is 0 or 1.

It also updates some checks that stop you from passing values resident
in series to routines that will hold onto that pointer over time.
For instance, you can't use a slot in a BLOCK! as the output location
for an evaluation, because the act of the evaluation might resize the
block and move the pointer.  The same applies to guarding REBVAL*.