Break out %sys-rebnod.h for REBVAL*/REBSER* superclass
The evolution of REBVAL and REBSER to share a common header happened
gradually over time, recently made more sophisticated by the bit
arrangement to distinguish from UTF-8.  This formally declares the
REBNOD* "Rebol Node" type as the superclass for both, even though it's
not a perfect name for it.  Flags common to both are then categorized
as NODE_FLAG_XXX.