Clearer split between series info bits and series header flags
The original R3-Alpha flags on series were 8 bits:

    SER_MARK = 1,  // Series was found during GC mark scan.
    SER_KEEP = 1<<1, // Series is permanent, do not GC it.
    SER_LOCK = 1<<2, // Series is locked, do not expand it
    SER_EXT  = 1<<3, // Series is external (library), do not GC it.
    SER_FREE = 1<<4, // mark series as removed
    SER_BARE = 1<<5, // Series has no links to GC-able values
    SER_PROT = 1<<6, // Series is protected from modification
    SER_MON  = 1<<7, // Monitoring

These flags were very scarce.  Ren-C made a tradeoff in growing the
size of a series node from 5x(void*) to 8x(void*) in order to implement
a trick of embedding arrays of size 1 directly into the node, as well
as other small series...plus other optimization tricks.  One of those
tricks made sure a flag set could do internal termination, at the cost
of 2 of the 32 used bits in the flag set.  Another trick reused flags
between value headers and series nodes to provide c common header.
(The word symbol table was also converted into being GC-able and
unified with the series nodes... pairing cells for API handles being
GC'able, etc. etc.)

One group of the new flags were in a field called "header" and another
in a field called "info".  There was not much logic in what would put
information in one place or the other.  Names like SERIES_FLAG_XXX and
REBSER_FLAG_XXX were not very clear, either.

This is a full going-over of the flags, splitting them into either
SERIES/ARRAY/CONTEXT_FLAG_XXX or SERIES_INFO_XXX.  A new concept is
to put the flags that tend to change over the lifetime of the series
into `info`, while the ones that are unlikely to change go in `header`.

Also, this takes flags out of enum {} declarations, because they are
using REBUPT as an unsigned integer of pointer size.  enum sizes and
types in C are not specified, so this uses the #define to get the
REBUPT constants from the macros directly.  At the same time it goes
ahead and makes it possible to distinguish UTF-8 characters from
internal END markers.

(It is still possible to make a change to distinguish free series nodes
from UTF-8, if that becomes interesting for the diagnostics, but it
would not be a "feature".)

This also makes the sacrifice of using full bytes for the series width
in the small series optimization as well as the datatype.  Those can
be read with a single operation on most architectures, and these are
extremely common values to be extracted.