Use C++11 or above in CMake
While technically possible to build Ren-C in C++98, this does not
actually give you any interesting additional checks (which rely on
things like <type_traits>). So there is not much point to building as
C++ over C unless C++11 is used.

This adds the flag to CMakeLists.txt to tell the compiler to use the
C++11 standard when building C++ files (however that is specified)