ANY and ALL chain "opt-outs" (return void)...add ANY? and ALL?
R3-Alpha's logic for handling "unset!" in ANY and ALL was to consider
it neither true nor false.  Yet still, it would overwrite the output
if it was in the last position.

This gave a somewhat unusual smattering of results:

    >> all [1 () 2]
    == 2

    >> all [1 2 ()]
    ;-- no value

    >> all []
    == true

    >> all [()]
    ;-- no value

Ren-C formalized this to where voids would "opt-out", yet would always
return a value.  This made ANY and ALL safe to use in conditionals.
Yet it still was throwing away information (`all [1 2 ()]` was true,
not 2).  The damage this did to ALL's ability to act as a chaining
operation wasn't a good "optimization", so this changes that.

Yet another problem in chaining was that the `any [()] => false` and
`any [()] => true` concept made it not possible to chain the idea of
opting-out.  These kinds of cases don't usually come up in conventional
coding, but when trying to compose operations they do...and the right
answer is to make them return void, so that they can chain.

For a friendlier version that guarantees it will work in conditionals,
this introduces the ANY? and ALL? variations.

    if any? [() ()] [
        print "this won't print...all opt-out means false"
    ]

    if all? [() ()] [
        print "this will print...all opt-out means true"
    ]