Remove memory poisoning
The intention of memory poisoning was to find memory problems. However,
because GC needs access to all of slots in the memory pool, including
the ones not allocated to any serieses, it needs to unpoison it before
accessing and repoison it again if it was poisoned, which adds some
complexity to the code. Moreover, because this is hidden behind a
compilation switch, it's not going through enough testing (even
compliation testing).

"always_malloc = 1" and address sanitizer/valgrind could be a good
alternative tool to find memory problems.

This removal was proposed by @Hostilefork.