Bring back sanitizer macros, remove ATTRIBUTE_NO_SANITIZE_ADDRESS
While complicating the code permanently with address sanitizer macros
is probably not worth it given the Always_Malloc option, it can be
useful to have the macros available for other debug scenarios.  This
brings them back, but then also removes the ATTRIBUTE_NO_SANITIZE
off of the gc functions since memory poisoning is not being used
anymore.