New loop return result policy, eliminate BREAK/RETURN
This enforces the concept that a loop will return BLANK! if it is
broken with BREAK, and that most loops will return the last body result
unless it is void or FALSE?, in which case the result will be coerced
to the true (but unusual-looking) result of BAR!.

By having a uniform policy, loops become easier to chain.  As a
simple example of how easy:

     for-both: function [
          'var [word!]
          a [block!]
          b [block!]
          body [block!]
     ][
          all [
              for-each var a body
              for-each var b body
          ]
     ]

Supporting BREAK/RETURN would undermine the ability to write such
compositions.  So instead, users are advised to use THROW and CATCH,
which is a generic mechanism for overriding return values, and one
could even argue a better looking one:

    while [x < 10] [break/return "result"]
    catch [while [x < 10] [throw "result"]]