Disallow `<opt>` on refinement arguments
A refinement argument can only be void if the refinement itself is
completely unused.  Hence it doesn't represent the same thing as
an ordinary void argument...from the callee's perspective, there is no
way to tell the difference between:

    append/dup [a b c] 'd ()
    append [a b c] 'd

So this triggers an error if `<opt>` is applied to a refinement
argument.