Improve AS_SERIES (takes over ARR_SERIES), AS_ARRAY, AS_CONTEXT
This unifies the AS_SERIES and ARR_SERIES, without compromising on
performance...since the type assertion is only added in the C++
build when AS_SERIES is used on a type not actually known to be
a series (e.g. a REBNOD* or a void*).

Also differentiates REBCTX* and REBSPC* for specifiers, to permit a
wider range of specifier usages...such as raw frame pointers, pairings,
or plain arrays.  These are hoped to be able to help reduce the tax
caused by reification for the garbage collector, including for
plain definitional returns.

Includes some tweaks for fixing clang warnings in trying to figure out
why this commit refuses to build on Travis.