Update Check_Memory and add to debug Init and Shutdown
Rebol's early development predates Valgrind and Address Sanitizer, so
R3-Alpha had a macro for checking memory at various moments times,
based on how paranoid a checking build you were running.  This meant
the CHECK_MEMORY() macro was scattered throughout the system anywhere
that presumably a bug had been found historically.

Ren-C relies more heavily on tooling, but Check_Memory can still do
some basic sanity checking on the memory pools that an external tool
can't do (although the Always_Malloc idea that doesn't use pools will
be a more thorough check for the same kinds of problems).

Still, it might be useful in *some* debug scenario.  However, since no
one had actually called it for some time it had atrophied and needed to
be brought up to date.

This does that update, kills the macro, eliminates the user-facing
error declaration (bad memory pools is a panic() condition, and is
not recoverable from a longjmp).  It also adds one call to the memory
check on init and shutdown in the debug build.  This way, the basic
sanity check it offers will stay in use and up to date, so it will
have at least some value.