Remove pi/blank from bootstrap constants, some simplification
In trying to minimize the unnecessary work during bootstrap, this
takes the definition of the constant "pi" out and moves it into the
math mezzanine.  It also takes away the word "blank", instead having
any early bootstrap code use the blank literal.  This helps focus
on the "required" definition of true and false, and why that is
currently necessary.

Also fixes a bug with initialization of command line arguments where
the created block's length was inconsistent with its end marker.
Plus removes some tabs in the %mezz-math.r file.