Kill "boot strings", begin RM_XXX "Rebol Messages", update Dump_XXX
Keeping "resource" strings in Rebol format--as opposed to in C--offers
several theoretical advantages.  One is that they are compressed along
with all the rest of the boot data.  Another is that they can in theory
be swapped out without touching the C code...and even changed at
runtime, thus permitting internationalization.  Also, if the string
needs to be a template and parameterized, it can be done as a dialect
in a way that is more graceful than the %d and %s format specifiers
of C strings.

However, technical limitations stop some strings from coming from
LOADed code.  One set of strings that cannot be handled this way is
strings required during LOAD itself...such as the names of months.
Others would be error messages that might need to be delivered so
early in the boot that LOAD hasn't run.

R3-Alpha created two categories of these strings.  One set were in
%a-constants.c...which was specifically strings that were needed
during load.  The second was in a Rebol file called %strings.r, which
would be LOADed as part of the boot block, and then compacted into
an array to be usable from C code without having to go through a
STRING! Rebol value to use them.

It didn't really make a lot of sense to have user-visible messages
which were extracted into C strings after the boot block had been
loaded.  That would be redundant with the values.  The strings that
made sense to hold in the boot block should be left there, but also
use block-structured templating (the way the errors do) instead of
rehashing the implementation of `printf()`.

So to pursue that simplification, this takes the remaining strings out
of %strings.r and moves them to %a-constants.c, along with a beginning
of defining the RM_XXX "Rebol messages" that would belong in some
%messages.r file from the boot block.  While some messages were
extracted, others that were debug-build-only using the printf format
could just use printf directly (since the debug build includes it).
Using printf exposed compiler warnings that exist for checking the
parameters of printf, so this addressed those...and then patched the
Dump_XXX routines which had fallen into disrepair.