Simplify Boot_Block inclusion to clarify dependencies
The file %boot.c contains the compressed data for all the Rebol boot
code, which includes everything from error message templates to
the list of datatypes and the mezzanine code.  The block has several
sections, which were aliased by a struct called Boot_Block.

Because Boot_Block was a global variable initialized by Load_Boot(),
it was not quite clear what parts of the boot process used what.  There
were also redundant checks of the boot block's integrity which could be
done when those sections of the boot block were being used.  This
factors out the sections to become explicit parameters to those
init routines that use them.  It also eliminates the %booters.r file
which is obsolete.