Move codec registration out of Init_Core, use service function
The idea of a "codec" was that of a specialized API that C code could
be written to (but not linked against any API that understood Rebol
values), which could encode and decode binary streams as Rebol types.

These codecs were written in C and built into the system, listed in
the system context.  As part of the core setup process that ran as
Rebol code, the list of codecs registered by the C was scanned and
correlated with file extensions.  This meant that any Register_Codec
calls from C that came later would not add the codec to the lists
(and that the file extensions weren't coming from the Register_Codec
call...)

This changes Register_Codec() the C call to transfer control to
REGISTER-CODEC*, a system Rebol call, so that there is not a "moment"
that the codecs are scanned which could be "missed".  As a result,
the codecs can be added after the Init_Core() has run.  This is a
first step toward making the codecs more dynamic.